+++ game/src/api/IGameState.hpp
@@ -16,7 +16,7 @@ void fromString(const std::string &str);
         std::vector<Player> players;
         bool private_session;
         bool rolled;
-        std::vector<double> rolls;
+        std::vector<int64_t> rolls;
         int64_t spectators;
         int64_t turn_index;
         std::vector<bool> used;
