+++ game/src/api/RollMsg.hpp
@@ -12,7 +12,7 @@ namespace API {
     struct RollMsg {
 std::string toString() const;
 void fromString(const std::string &str);
-        std::vector<double> rolls;
+        std::vector<int64_t> rolls;
         RollMsgType type = static_cast<RollMsgType>(0);
     };
 }
